# API/openapi/API_v2.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Mangette v2",
    "version": "2.0"
  },
  "paths": {
...
    "/v2/MetadataFetcher/Links": {
      "get": {
        "tags": [
          "MetadataFetcher"
        ],
        "summary": "Returns all API.Schema.MangaContext.MetadataFetchers.MetadataEntry",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataEntry"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation"
          }
        }
      }
    },
    "/v2/MetadataFetcher/Links/{MangaId}": {
      "get": {
        "tags": [
          "MetadataFetcher"
        ],
        "summary": "Returns all API.Schema.MangaContext.MetadataFetchers.MetadataEntry for API.Schema.MangaContext.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataEntry"
                  }
                }
              }
            }
          },
          "404": {
            "description": "API.Schema.MangaContext.Manga with MangaId not found",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v2/MetadataFetcher/{MetadataFetcherName}/SearchManga/{MangaId}": {
      "post": {
        "tags": [
          "MetadataFetcher"
        ],
        "summary": "Searches API.Schema.MangaContext.MetadataFetchers.MetadataFetcher (Metadata-Sites) for Manga-Metadata",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MetadataFetcherName",
            "in": "path",
            "description": "API.Schema.MangaContext.MetadataFetchers.MetadataFetcher.Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Instead of using the MangaId for search on Website, use a specific term",
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataSearchResult"
                  }
                }
              }
            }
          },
          "400": {
            "description": "API.Schema.MangaContext.MetadataFetchers.MetadataFetcher (Metadata-Sites) with MetadataFetcherName does not exist",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "API.Schema.MangaContext.Manga with MangaId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v2/MetadataFetcher/{MetadataFetcherName}/Link/{MangaId}": {
      "post": {
        "tags": [
          "MetadataFetcher"
        ],
        "summary": "Links API.Schema.MangaContext.MetadataFetchers.MetadataFetcher (Metadata-Sites) using Provider-Specific Identifier to API.Schema.MangaContext.Manga",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "description": "API.Schema.MangaContext.Manga.Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MetadataFetcherName",
            "in": "path",
            "description": "API.Schema.MangaContext.MetadataFetchers.MetadataFetcher.Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "!:MetadataFetcherName-Specific ID",
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/MetadataEntry"
                }
              }
            }
          },
          "400": {
            "description": "API.Schema.MangaContext.MetadataFetchers.MetadataFetcher (Metadata-Sites) with MetadataFetcherName does not exist",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "API.Schema.MangaContext.Manga with MangaId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v2/MetadataFetcher/{MetadataFetcherName}/Unlink/{MangaId}": {
      "post": {
        "tags": [
          "MetadataFetcher"
        ],
        "summary": "Un-Links API.Schema.MangaContext.MetadataFetchers.MetadataFetcher (Metadata-Sites) from API.Schema.MangaContext.Manga",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MetadataFetcherName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "API.Schema.MangaContext.MetadataFetchers.MetadataFetcher (Metadata-Sites) with MetadataFetcherName does not exist",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "API.Schema.MangaContext.Manga with MangaId not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error during Database Operation",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "412": {
            "description": "No API.Schema.MangaContext.MetadataFetchers.MetadataEntry linking API.Schema.MangaContext.Manga and API.Schema.MangaContext.MetadataFetchers.MetadataFetcher found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v2/MetadataFetcher/{MetadataFetcherName}/Update/{MangaId}": {
      "post": {
        "tags": [
          "MetadataFetcher"
        ],
        "summary": "Updates the Metadata of API.Schema.MangaContext.Manga with MangaId",
        "parameters": [
          {
            "name": "MangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MetadataFetcherName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata updated"
          },
          "400": {
            "description": "API.Schema.MangaContext.MetadataFetchers.MetadataFetcher not found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "No API.Schema.MangaContext.MetadataFetchers.MetadataEntry linking API.Schema.MangaContext.Manga and API.Schema.MangaContext.MetadataFetchers.MetadataFetcher found",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MetadataEntry": {
        "type": "object",
        "properties": {
          "mangaId": {
            "type": "string",
            "nullable": true
          },
          "metadataFetcherName": {
            "type": "string",
            "nullable": true
          },
          "identifier": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MetadataSearchResult": {
        "required": [
          "coverUrl",
          "description",
          "identifier",
          "name",
          "url"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "minLength": 1,
            "type": "string",
            "description": "Identifier specific to the MetadataFetcher"
          },
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the Manga"
          },
          "url": {
            "minLength": 1,
            "type": "string",
            "description": "Url to the result"
          },
          "description": {
            "minLength": 1,
            "type": "string",
            "description": "Description of the Manga"
          },
          "coverUrl": {
            "minLength": 1,
            "type": "string",
            "description": "Url to the cover if available"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}